# faster sharing-aware cost extraction