--- a/src/extract/mod.rs
+++ b/src/extract/mod.rs
@@ -4,6 +4,7 @@ pub use crate::*;
 
 pub mod bottom_up;
 pub mod greedy_dag;
+pub mod greedy_dag_1;
 
 #[cfg(feature = "ilp-cbc")]
 pub mod ilp_cbc;