@TrevorHansen 2023-09-13
> Oh interesting. I think this is implying that there are a couple heavy hitters in terms of time spent in extraction. The geometric mean reports that this code is about as fast, but in absolute time it's much faster. Any idea what's going on there?


Yes, Max, you're right about that. With the 211 inputs, these 4 problems provide about 95% of the time improvement: 

| file                                  | faster greedy dag extractor (ms) | old extractor with std hashmap (ms) |
|---------------------------------|---------------------------|--------------------------|
|data/tensat/nasrnn_acyclic.json | 312 | 2,560|
|data/tensat/nasneta.json | 1,392 | 15,290|
|data/tensat/nasneta_acyclic.json | 2,602 | 17,613|
|data/tensat/nasrnn.json | 3,163 | 8,904|
|---------------------------------|---------------------------|--------------------------|
| total | 7,469 |  44,367|

The other 207 problems take 1.3 seconds to be solved by the new extractor and 2.8 seconds for the old extractor. So there are a lot more easy problems than hard problems, which hides the improvement in the geometric average.